requirements.txt
flask>=2.3.0
werkzeug>=2.3.0
torch>=2.2.0
torchaudio>=2.2.0
nemo_toolkit[all]>=1.18.0
numpy>=1.26.0
librosa>=0.10.0
soundfile>=0.12.0